fix(twland): send wl_keyboard.keymap fd via SCM_RIGHTS
The keymap event's fd argument was being serialized into the payload as a
u32, corrupting the wire format and disconnecting every libwayland client
that bound a keyboard. The root cause was that the send path had no
SCM_RIGHTS fd-passing capability at all.

Extract a wire-protocol layer (src/wire.rs) owning message framing, the
argument codec, and SCM_RIGHTS fd passing. Add send_message_with_fds and
create_empty_memfd (backed by Twilight's memfd_create syscall). Fix
send_keyboard_keymap to emit the correct 8-byte (format, size) payload
with a real empty memfd attached out-of-band.

Remove all duplicated wire/codec code from main.rs.

Closes #39
